refactor: 일-토 라벨 컴포넌트화
