+++ src/screens/OppuScreen/components/WeekDayLabels/styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  weekDayLabelRow: {
+    flexDirection: 'row',
+  },
+  weekDayLabelCell: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 8,
+  },
+});
+
+export default styles;
